Fix array bounds issue in pittpatt wrapper
The verions of janus_create_template which takes pre-defined bounding boxes as
input is implemented in the pp5 wrapper by cropping the bounding boxes, then
running detection.

These (different sized) crops were stored on the same janus_media, which is
supposed to be for same size images only. So, of a set of N crops, only the last
one added was stored with the correct images dimensions.

This could (rarely) cause seg-faults, and should generally just mess up crops
of images, except for the last one added to the media (which is fine).